requirements.txt
aiohttp
httpx
pillow
piexif
websockets
tiktoken
pydantic
regex
requests
google-genai
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery
numpy
onnxruntime
opencv-python
mediapipe
huggingface-hub
